Refactor Skill and Spot controllers to accept DTOs, fix mapped path
