+++ employee-rostering-backend/src/main/java/org/optaweb/employeerostering/domain/skill/Skill.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaweb.employeerostering.domain;
+package org.optaweb.employeerostering.domain.skill;
 
 import javax.persistence.Entity;
 import javax.persistence.Table;
@@ -23,6 +23,8 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
+import org.optaweb.employeerostering.domain.AbstractPersistable;
+
 @Entity
 @Table(uniqueConstraints = @UniqueConstraint(columnNames = {"tenantId", "name"}))
 public class Skill extends AbstractPersistable {
@@ -30,7 +32,7 @@ public class Skill extends AbstractPersistable {
     @NotNull
     @Size(min = 1, max = 120)
     @Pattern(regexp = "^(?!\\s).*(?<!\\s)$", message = "Name should not contain any leading or trailing whitespaces")
-    private String name;
+    protected String name;
 
     @SuppressWarnings("unused")
     public Skill() {
