+++ employee-rostering-backend/src/test/java/org/optaweb/employeerostering/controller/SpotControllerTest.java
@@ -26,8 +26,8 @@ import javax.transaction.Transactional;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.optaweb.employeerostering.domain.Skill;
-import org.optaweb.employeerostering.domain.Spot;
+import org.optaweb.employeerostering.domain.skill.Skill;
+import org.optaweb.employeerostering.domain.spot.Spot;
 import org.optaweb.employeerostering.service.SkillService;
 import org.optaweb.employeerostering.service.SpotService;
 import org.slf4j.Logger;
